+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 * #186 Exp.in, Exp.notIn, Series.in, Series.notIn
 * #187 Exp.last(), Series.last()
 * #189 ExcelLoader: Duplicate empty column names
+* #190 DateTimeExp - direct converter to DateExp and TimeExp
 
 ## Release 0.17